refactor: domain의 model에서 toDto 메소드 제거
- domain의은 data의 dto를 몰라야한다고 정의
   - 따라서, domain에서 model이 dto로 변환되는 함수를 data entity로 이동시킴
